Move dolphin OS headers to subdirectory
