+++ include/dolphin/os/OSAlarm.h
@@ -1,8 +1,8 @@
-#ifndef _DOLPHIN_OSALARM_H_
-#define _DOLPHIN_OSALARM_H_
+#ifndef _DOLPHIN_OS_OSALARM_H_
+#define _DOLPHIN_OS_OSALARM_H_
 
-#include "dolphin/OSContext.h"
 #include "dolphin/os.h"
+#include "dolphin/os/OSContext.h"
 
 typedef struct OSAlarm OSAlarm;
 typedef void (*OSAlarmHandler)(OSAlarm* alarm, OSContext* context);
@@ -22,4 +22,4 @@ void OSCreateAlarm(OSAlarm* alarm);
 void OSSetAlarm(OSAlarm* alarm, OSTime tick, OSAlarmHandler handler);
 void OSCancelAlarm(OSAlarm* alarm);
 
-#endif // _DOLPHIN_OSALARM_H_
+#endif
