+++ include/dolphin/os/OSAlloc.h
@@ -1,5 +1,5 @@
-#ifndef _DOLPHIN_OSALLOC_H_
-#define _DOLPHIN_OSALLOC_H_
+#ifndef _DOLPHIN_OS_OSALLOC_H_
+#define _DOLPHIN_OS_OSALLOC_H_
 
 #include "dolphin/types.h"
