+++ include/dolphin/os/OSFont.h
@@ -1,5 +1,5 @@
-#ifndef _DOLPHIN_OSFONT_H_
-#define _DOLPHIN_OSFONT_H_
+#ifndef _DOLPHIN_OS_OSFONT_H_
+#define _DOLPHIN_OS_OSFONT_H_
 
 #include "dolphin/types.h"
