+++ include/dolphin/os/OSMutex.h
@@ -1,7 +1,7 @@
-#ifndef _DOLPHIN_OSMUTEX_H_
-#define _DOLPHIN_OSMUTEX_H_
+#ifndef _DOLPHIN_OS_OSMUTEX_H_
+#define _DOLPHIN_OS_OSMUTEX_H_
 
-#include "dolphin/OSThread.h"
+#include "dolphin/os/OSThread.h"
 
 typedef struct OSMutex {
     /*0x00*/ OSThreadQueue queue;
