+++ include/dolphin/os/OSReset.h
@@ -1,5 +1,5 @@
-#ifndef _DOLPHIN_OSRESET_H_
-#define _DOLPHIN_OSRESET_H_
+#ifndef _DOLPHIN_OS_OSRESET_H_
+#define _DOLPHIN_OS_OSRESET_H_
 
 #define OS_RESET_RESTART 0
 #define OS_RESET_HOTRESET 1
