+++ include/dolphin/os/OSResetSW.h
@@ -1,8 +1,8 @@
-#ifndef _DOLPHIN_OSRESETSW_H_
-#define _DOLPHIN_OSRESETSW_H_
+#ifndef _DOLPHIN_OS_OSRESETSW_H_
+#define _DOLPHIN_OS_OSRESETSW_H_
 
-#include "dolphin/types.h"
 #include "dolphin/hw_regs.h"
+#include "dolphin/types.h"
 
 void __OSResetSWInterruptHandler();
 void (*OSSetResetCallback(void (*callback)()))();
