+++ include/dolphin/os/OSSerial.h
@@ -57,4 +57,4 @@ void SIGetResponse(long chan, void* data);
 int SITransfer(long chan, void* output, unsigned long outputBytes, void* input, unsigned long inputBytes,
                void (*callback)(long, unsigned long, struct OSContext*), long long time);
 
-#endif // _DOLPHIN_OSSERIAL_H
+#endif
